requirements.txt
attrs==21.4.0
black==22.6.0
cachelib==0.9.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
decorator==4.4.2
Flask==2.1.3
flask-expects-json==1.7.0
idna==3.3
imageio==2.19.5
imageio-ffmpeg==0.4.7
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.7.2
MarkupSafe==2.1.1
moviepy==1.0.3
mypy-extensions==0.4.3
numpy==1.23.1
pathspec==0.9.0
Pillow==9.2.0
platformdirs==2.5.2
proglog==0.1.10
psutil==5.9.1
pyrsistent==0.18.1
python-magic==0.4.27
requests==2.28.1
tomli==2.0.1
tqdm==4.64.0
urllib3==1.26.10
Werkzeug==2.1.2
